pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kalibr"
version = "1.1.0"
description = "Unified LLM Observability & Multi-Model AI Integration Framework - Deploy to GPT, Claude, Gemini, Copilot with full telemetry"
authors = [{name = "Kalibr Team", email = "team@kalibr.dev"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = [
    "ai", "mcp", "gpt", "claude", "gemini", "copilot",
    "openai", "anthropic", "google", "microsoft",
    "observability", "telemetry", "tracing", "llm",
    "schema-generation", "api", "multi-model"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.27.0",
    "tiktoken>=0.8.0",
    "fastapi>=0.110.1",
    "uvicorn>=0.25.0",
    "pydantic>=2.6.4",
    "typer>=0.9.0",
    "python-multipart>=0.0.9",
    "rich>=10.0.0",
    "requests>=2.31.0",
    # OpenTelemetry for SDK instrumentation (Phase 1)
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
kalibr = "kalibr.cli.main:app"

[project.urls]
Homepage = "https://github.com/devonakelley/kalibr-sdk"
Documentation = "https://github.com/devonakelley/kalibr-sdk#readme"
Repository = "https://github.com/devonakelley/kalibr-sdk"
Issues = "https://github.com/devonakelley/kalibr-sdk/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["kalibr*"]
exclude = ["tests*", "examples*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
